Added clang-format with LLVM style. Formatted all sources.

+++ include/ibex/Function.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <ibex/Storage.h>
+
 #include <functional>
 
 namespace ibex {
@@ -9,8 +10,8 @@ template <typename, size_t>
 class Function;
 
 template <size_t Size, typename Result, typename... Arguments>
-class Function<Result (Arguments...), Size> final {
-private:
+class Function<Result(Arguments...), Size> final {
+ private:
   // ---------------------------------------------------------------------------
   // Types
   // ---------------------------------------------------------------------------
@@ -18,7 +19,7 @@ private:
   struct ErasedFunctorHolder {
     virtual ~ErasedFunctorHolder() {}
     virtual ReturnType operator()(Args&&...) = 0;
-    virtual void moveInto (void*) = 0;
+    virtual void moveInto(void*) = 0;
   };
 
   template <typename Functor, typename ReturnType, typename... Args>
@@ -26,75 +27,68 @@ private:
     // Non-erased Functor
     Functor f;
 
-    FunctorHolder (Functor&& func) : f (std::move(func)) {}
-    ~FunctorHolder () override = default;
+    FunctorHolder(Functor&& func) : f(std::move(func)) {}
+    ~FunctorHolder() override = default;
 
     ReturnType operator()(Args&&... args) override {
-      return f (std::forward<Arguments> (args)...);
+      return f(std::forward<Arguments>(args)...);
     }
 
-    void moveInto (void* destination) override {
-      new (destination) FunctorHolder (std::move(f));
+    void moveInto(void* destination) override {
+      new (destination) FunctorHolder(std::move(f));
     }
   };
 
   using holder_t = ErasedFunctorHolder<Result, Arguments...>;
-  
-  
+
   // ---------------------------------------------------------------------------
   // Members
   // ---------------------------------------------------------------------------
-  bool m_isValid{false}; //
+  bool m_isValid{false};  //
   mutable PolyStorage<holder_t, Size> m_storage;
-  
+
   // ---------------------------------------------------------------------------
   // Public Functions
   // ---------------------------------------------------------------------------
-public:
-
+ public:
   // Create an empty function
   Function() = default;
-  
+
   ~Function() {
-    if (m_isValid)
-      m_storage.destroy();
+    if (m_isValid) m_storage.destroy();
   }
-   
+
   // Move CTOR
-  Function (Function&& other) {
-    moveFrom(std::move(other));
-  }
-  
+  Function(Function&& other) { moveFrom(std::move(other)); }
+
   template <typename Functor>
-  Function (Functor&& f) {
+  Function(Functor&& f) {
     using exactFunctionHolder_t = FunctorHolder<Functor, Result, Arguments...>;
     m_storage.template create<exactFunctionHolder_t>(std::move(f));
     m_isValid = true;
   }
 
-  Function& operator= (Function&& other) {
+  Function& operator=(Function&& other) {
     if (this != &other) {
       moveFrom(std::move(other));
     }
     return *this;
   }
-  
+
   // Invoke the contained target.
   // Note: Undefined behaviour if no valid target is stored .
-  Result operator() (Arguments&&... args) const {
+  Result operator()(Arguments&&... args) const {
     if (!m_isValid) throw std::bad_function_call{};
-    return m_storage.get()(std::forward<Arguments> (args)...);
+    return m_storage.get()(std::forward<Arguments>(args)...);
   }
-  
+
   // Check whether a valid function is stored
-  operator bool() const {
-    return m_isValid;
-  }
-  
+  operator bool() const { return m_isValid; }
+
   // ---------------------------------------------------------------------------
   // Private Functions
   // ---------------------------------------------------------------------------
-private:
+ private:
   void moveFrom(Function&& other) {
     clear();
     if (other.m_isValid) {
@@ -103,14 +97,13 @@ private:
       other.clear();
     }
   }
-  
+
   void clear() {
     if (m_isValid) {
       m_storage.destroy();
       m_isValid = false;
     }
   }
-  
 };
 
-}
+}  // namespace ibex
